refactor(Connector): replace ConnectorCard with ConnectorDisplay
- Updated generateConnectorStories script to use ConnectorDisplay instead of ConnectorCard for story generation.
- Removed ConnectorCard component and its associated stories, transitioning to ConnectorDisplay for improved consistency.
- Adjusted imports in various components to reflect the change from ConnectorCard to ConnectorDisplay, ensuring all references are updated accordingly.
