--- packages/ui-v1/__stories__/ConnectorCard.stories.tsx
@@ -1,255 +1,0 @@
-// generated file. Do not modify by hand
-
-// Generated by generateConnectorStories.ts
-import type {Meta, StoryObj} from '@storybook/react'
-import type {ConnectorName} from '@openint/all-connectors/name'
-
-import {defConnectors} from '@openint/all-connectors/connectors.def'
-import {getConnectorModel} from '@openint/api-v1/models'
-import {ConnectorCard} from '../domain-components/ConnectorCard'
-
-function StoryWrapper(props: {name: ConnectorName}) {
-  const def = defConnectors[props.name]
-  if (!def) {
-    throw new Error('Connector ' + props.name + ' not found')
-  }
-  const connector = getConnectorModel(def, {
-    includeSchemas: true,
-  })
-
-  return <ConnectorCard connector={connector} />
-}
-
-const meta: Meta<typeof StoryWrapper> = {
-  title: 'All Connectors/connector',
-  component: StoryWrapper,
-  parameters: {layout: 'centered'},
-  tags: ['autodocs'],
-}
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const acmeOauth2ConnectorCard: Story = {
-  args: {name: 'acme-oauth2'},
-}
-
-export const aircallConnectorCard: Story = {
-  args: {name: 'aircall'},
-}
-
-export const confluenceConnectorCard: Story = {
-  args: {name: 'confluence'},
-}
-
-export const discordConnectorCard: Story = {
-  args: {name: 'discord'},
-}
-
-export const facebookConnectorCard: Story = {
-  args: {name: 'facebook'},
-}
-
-export const githubConnectorCard: Story = {
-  args: {name: 'github'},
-}
-
-export const gongConnectorCard: Story = {
-  args: {name: 'gong'},
-}
-
-export const googleCalendarConnectorCard: Story = {
-  args: {name: 'google-calendar'},
-}
-
-export const googleDocsConnectorCard: Story = {
-  args: {name: 'google-docs'},
-}
-
-export const googleDriveConnectorCard: Story = {
-  args: {name: 'google-drive'},
-}
-
-export const googleMailConnectorCard: Story = {
-  args: {name: 'google-mail'},
-}
-
-export const googleSheetConnectorCard: Story = {
-  args: {name: 'google-sheet'},
-}
-
-export const hubspotConnectorCard: Story = {
-  args: {name: 'hubspot'},
-}
-
-export const instagramConnectorCard: Story = {
-  args: {name: 'instagram'},
-}
-
-export const intercomConnectorCard: Story = {
-  args: {name: 'intercom'},
-}
-
-export const jiraConnectorCard: Story = {
-  args: {name: 'jira'},
-}
-
-export const leverConnectorCard: Story = {
-  args: {name: 'lever'},
-}
-
-export const linearConnectorCard: Story = {
-  args: {name: 'linear'},
-}
-
-export const linkedinConnectorCard: Story = {
-  args: {name: 'linkedin'},
-}
-
-export const notionConnectorCard: Story = {
-  args: {name: 'notion'},
-}
-
-export const outreachConnectorCard: Story = {
-  args: {name: 'outreach'},
-}
-
-export const pipedriveConnectorCard: Story = {
-  args: {name: 'pipedrive'},
-}
-
-export const quickbooksConnectorCard: Story = {
-  args: {name: 'quickbooks'},
-}
-
-export const redditConnectorCard: Story = {
-  args: {name: 'reddit'},
-}
-
-export const salesloftConnectorCard: Story = {
-  args: {name: 'salesloft'},
-}
-
-export const sharepointConnectorCard: Story = {
-  args: {name: 'sharepoint'},
-}
-
-export const slackConnectorCard: Story = {
-  args: {name: 'slack'},
-}
-
-export const twitterConnectorCard: Story = {
-  args: {name: 'twitter'},
-}
-
-export const xeroConnectorCard: Story = {
-  args: {name: 'xero'},
-}
-
-export const zohoDeskConnectorCard: Story = {
-  args: {name: 'zoho-desk'},
-}
-
-export const airtableConnectorCard: Story = {
-  args: {name: 'airtable'},
-}
-
-export const apolloConnectorCard: Story = {
-  args: {name: 'apollo'},
-}
-
-export const brexConnectorCard: Story = {
-  args: {name: 'brex'},
-}
-
-export const codaConnectorCard: Story = {
-  args: {name: 'coda'},
-}
-
-export const finchConnectorCard: Story = {
-  args: {name: 'finch'},
-}
-
-export const firebaseConnectorCard: Story = {
-  args: {name: 'firebase'},
-}
-
-export const foreceiptConnectorCard: Story = {
-  args: {name: 'foreceipt'},
-}
-
-export const greenhouseConnectorCard: Story = {
-  args: {name: 'greenhouse'},
-}
-
-export const heronConnectorCard: Story = {
-  args: {name: 'heron'},
-}
-
-export const lunchmoneyConnectorCard: Story = {
-  args: {name: 'lunchmoney'},
-}
-
-export const mercuryConnectorCard: Story = {
-  args: {name: 'mercury'},
-}
-
-export const mergeConnectorCard: Story = {
-  args: {name: 'merge'},
-}
-
-export const mootaConnectorCard: Story = {
-  args: {name: 'moota'},
-}
-
-export const onebrickConnectorCard: Story = {
-  args: {name: 'onebrick'},
-}
-
-export const plaidConnectorCard: Story = {
-  args: {name: 'plaid'},
-}
-
-export const postgresConnectorCard: Story = {
-  args: {name: 'postgres'},
-}
-
-export const rampConnectorCard: Story = {
-  args: {name: 'ramp'},
-}
-
-export const saltedgeConnectorCard: Story = {
-  args: {name: 'saltedge'},
-}
-
-export const splitwiseConnectorCard: Story = {
-  args: {name: 'splitwise'},
-}
-
-export const stripeConnectorCard: Story = {
-  args: {name: 'stripe'},
-}
-
-export const tellerConnectorCard: Story = {
-  args: {name: 'teller'},
-}
-
-export const togglConnectorCard: Story = {
-  args: {name: 'toggl'},
-}
-
-export const twentyConnectorCard: Story = {
-  args: {name: 'twenty'},
-}
-
-export const venmoConnectorCard: Story = {
-  args: {name: 'venmo'},
-}
-
-export const wiseConnectorCard: Story = {
-  args: {name: 'wise'},
-}
-
-export const yodleeConnectorCard: Story = {
-  args: {name: 'yodlee'},
-}
