+++ packages/ui-v1/components/schema-form/fields.tsx
@@ -6,7 +6,7 @@ import Image from 'next/image'
 import {useState} from 'react'
 import {env} from '@openint/env'
 import {Input, Switch} from '@openint/shadcn/ui'
-import {ConnectorBadges} from '../../domain-components/ConnectorCard'
+import {ConnectorBadges} from '../../domain-components/ConnectorDisplay'
 import {ConnectorScopes} from '../ConnectorScopes'
 import {CopyID} from '../CopyID'
 import {CredentialsField} from './fields/CredentialsField'
