+++ packages/ui-v1/domain-components/__stories__/ConnectorBadges.stories.tsx
@@ -1,6 +1,6 @@
 import type {Meta, StoryObj} from '@storybook/react'
 
-import {ConnectorBadges} from '../ConnectorCard'
+import {ConnectorBadges} from '../ConnectorDisplay'
 import {FIXTURES} from './fixtures'
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
